+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Improvements
 
+* `Browser` now has a `redirectAfter` method that redirects to the given URL after the given number of seconds.
 * `Menu` now outputs semantic HTML. Links are rendered in HTML lists, with submenus rendered with nested lists.
 * `CustomDragLayer` component can now take className prop.